+++ example_site/src/constants/app.js
@@ -58,31 +58,6 @@ export const partnerConfigurations = {
     [partnerPresets.spectrum]: spectrumConfiguration,
 };
 
-export const months = [
-    'January',
-    'February',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-    'August',
-    'September',
-    'October',
-    'November',
-    'December'
-];
-
-export const days = [
-    'Sunday',
-    'Monday',
-    'Tuesday',
-    'Wednesday',
-    'Thursday',
-    'Friday',
-    'Saturday'
-];
-
 export const accountVideos = [
     {
         videoName: "Bold",
